refactor(services): split cell/dispatch.rs into submodules
Extract constants re-exports, dispatch router, name lookup, and tests into
separate files under crates/services/src/cell/dispatch/.

Public API unchanged. No behavior changes.
