+++ crates/services/src/cell/dispatch/names.rs
@@ -1,0 +1,131 @@
+//! Human-readable names for cell method indices (used in logging).
+
+use super::constants::*;
+
+/// Return a human-readable name for a cell method index, if known.
+pub fn cell_method_name(index: u16) -> &'static str {
+    match index {
+        // SGWBeing
+        CM_SET_TARGET_ID => "setTargetID",
+        CM_SET_MOVEMENT_TYPE => "setMovementType",
+        // SGWAbilityManager
+        CM_TOGGLE_COMBAT_DEBUG => "toggleCombatDebug",
+        CM_TOGGLE_COMBAT_VERBOSE_DEBUG => "toggleCombatVerboseDebug",
+        CM_CONFIRMATION_RESPONSE => "confirmationResponse",
+        // SGWCombatant
+        CM_SET_CROUCHED => "setCrouched",
+        CM_TOGGLE_HEAL_DEBUG => "toggleHealDebug",
+        CM_REQUEST_HOLSTER_WEAPON => "requestHolsterWeapon",
+        // OrganizationMember
+        CM_ORG_INVITE_RESPONSE => "organizationInviteResponse",
+        CM_ORG_LEAVE => "organizationLeave",
+        CM_ORG_BROADCAST_MINIMAP_PING => "BroadcastMinimapPing",
+        CM_ORG_STRIKE_TEAM_RESPONSE => "strikeTeamResponse",
+        CM_ORG_PVP_LEAVE_RESPONSE => "pvpOrganizationLeaveResponse",
+        CM_ORG_MOTD => "organizationMOTD",
+        CM_ORG_NOTE => "organizationNote",
+        CM_ORG_OFFICER_NOTE => "organizationOfficerNote",
+        CM_ORG_SET_RANK_PERMISSIONS => "organizationSetRankPermissions",
+        CM_ORG_SET_RANK_NAME => "organizationSetRankName",
+        CM_ORG_SQUAD_SET_LOOT_MODE => "squadSetLootMode",
+        CM_ORG_TRANSFER_CASH => "organizationTransferCash",
+        // MinigamePlayer
+        CM_MG_DEBUG_START => "debugStartMinigame",
+        CM_MG_DEBUG_SPECTATE => "debugSpectateMinigame",
+        CM_MG_DEBUG_JOIN => "debugJoinMinigame",
+        CM_MG_DEBUG_INSTANCE => "debugMinigameInstance",
+        CM_MG_START => "startMinigame",
+        CM_MG_END_CURRENT => "endCurrentMinigame",
+        CM_MG_REQUEST_SPECTATE_LIST => "requestSpectateList",
+        CM_MG_SPECTATE => "spectateMinigame",
+        CM_MG_REGISTER_HELP => "registerToMinigameHelp",
+        CM_MG_UPDATE_REGISTER_HELP => "updateRegisterToMinigameHelp",
+        CM_MG_START_CANCEL => "minigameStartCancel",
+        CM_MG_CALL_ACCEPT => "minigameCallAccept",
+        CM_MG_CALL_DECLINE => "minigameCallDecline",
+        CM_MG_CALL_ABORT => "minigameCallAbort",
+        CM_MG_CONTACT_REQUEST => "minigameContactRequest",
+        // GateTravel
+        CM_ON_DIAL_GATE => "onDialGate",
+        // SGWInventoryManager
+        CM_REMOVE_ITEM => "removeItem",
+        CM_LIST_ITEMS => "listItems",
+        CM_MOVE_ITEM => "moveItem",
+        CM_USE_ITEM => "useItem",
+        CM_REPAIR_ITEM_REQUEST => "repairItemRequest",
+        CM_REQUEST_ACTIVE_SLOT_CHANGE => "requestActiveSlotChange",
+        CM_REQUEST_AMMO_CHANGE => "requestAmmoChange",
+        // SGWMailManager
+        CM_REQUEST_MAIL_HEADERS => "requestMailHeaders",
+        CM_SEND_MAIL_MESSAGE => "sendMailMessage",
+        CM_ARCHIVE_MAIL_MESSAGE => "archiveMailMessage",
+        CM_DELETE_MAIL_MESSAGE => "deleteMailMessage",
+        CM_RETURN_MAIL_MESSAGE => "returnMailMessage",
+        CM_REQUEST_MAIL_BODY => "requestMailBody",
+        CM_TAKE_CASH_FROM_MAIL => "takeCashFromMailMessage",
+        CM_TAKE_ITEM_FROM_MAIL => "takeItemFromMailMessage",
+        CM_PAY_COD_FOR_MAIL => "payCODForMailMessage",
+        // Missionary
+        CM_ABANDON_MISSION => "abandonMission",
+        CM_SHARE_MISSION => "shareMission",
+        CM_SHARE_MISSION_RESPONSE => "shareMissionResponse",
+        // ContactListManager
+        CM_CONTACT_LIST_CREATE => "contactListCreate",
+        CM_CONTACT_LIST_DELETE => "contactListDelete",
+        CM_CONTACT_LIST_RENAME => "contactListRename",
+        CM_CONTACT_LIST_FLAGS_UPDATE => "contactListFlagsUpdate",
+        CM_CONTACT_LIST_ADD_MEMBERS => "contactListAddMembers",
+        CM_CONTACT_LIST_REMOVE_MEMBERS => "contactListRemoveMembers",
+        // SGWBlackMarketManager
+        CM_BM_SEARCH => "BMSearch",
+        CM_BM_CREATE_AUCTION => "BMCreateAuction",
+        CM_BM_PLACE_BID => "BMPlaceBid",
+        CM_BM_CANCEL_AUCTION => "BMCancelAuction",
+        CM_BM_START_WATCHING => "BMStartWatchingItem",
+        CM_BM_STOP_WATCHING => "BMStopWatchingItem",
+        // SGWPlayer own
+        CM_CALL_FOR_AID => "callForAid",
+        CM_USE_ABILITY => "useAbility",
+        CM_USE_ABILITY_ON_GROUND => "useAbilityOnGroundTarget",
+        CM_RESPAWN => "respawn",
+        CM_UNSTUCK => "unstuck",
+        CM_RESET_MY_ABILITIES => "resetMyAbilities",
+        CM_WHO => "who",
+        CM_INTERACT => "interact",
+        CM_DIALOG_BUTTON_CHOICE => "dialogButtonChoice",
+        CM_INITIAL_RESPONSE => "initialResponse",
+        CM_TRAIN_ABILITY => "trainAbility",
+        CM_PURCHASE_ITEMS => "purchaseItems",
+        CM_SELL_ITEMS => "sellItems",
+        CM_BUYBACK_ITEMS => "buybackItems",
+        CM_REPAIR_ITEMS => "repairItems",
+        CM_RECHARGE_ITEMS => "rechargeItems",
+        CM_SET_AUTO_CYCLE => "setAutoCycle",
+        CM_LOOT_ITEM => "lootItem",
+        CM_TRIGGER_REGION => "triggerClientHintedGenericRegion",
+        CM_REQUEST_RELOAD => "requestReload",
+        CM_CHOSEN_REWARDS => "chosenRewards",
+        CM_PET_INVOKE_ABILITY => "petInvokeAbility",
+        CM_PET_ABILITY_TOGGLE => "petAbilityToggle",
+        CM_PET_CHANGE_STANCE => "petChangeStance",
+        CM_SET_RING_TRANSPORTER_DEST => "setRingTransporterDestination",
+        CM_WORLD_INSTANCE_RESET => "onWorldInstanceReset",
+        CM_UPDATE_SYSTEM_OPTIONS => "updateSystemOptions",
+        CM_ORG_CREATION => "onOrganizationCreation",
+        CM_SPEND_APPLIED_SCIENCE_POINTS => "spendAppliedSciencePoints",
+        CM_CRAFT => "craft",
+        CM_RESEARCH => "research",
+        CM_REVERSE_ENGINEER => "reverseEngineer",
+        CM_ALLOYING => "alloying",
+        CM_RESPEC_CRAFTING => "respecCrafting",
+        CM_CLIENT_CHALLENGE_RESPONSE => "onClientChallengeResponse",
+        CM_SEND_DUEL_RESPONSE => "sendDuelResponse",
+        CM_DUEL_FORFEIT => "duelForfeit",
+        CM_TRADE_REQUEST => "tradeRequest",
+        CM_TRADE_REQUEST_CANCEL => "tradeRequestCancel",
+        CM_TRADE_UPDATE_PROPOSAL => "tradeUpdateProposal",
+        CM_TRADE_LOCK_STATE => "tradeLockState",
+        CM_CANCEL_MOVIE => "cancelMovie",
+        _ => "unknown",
+    }
+}
